[INTERNAL][E] Skip non-shared resources for ProjectDeltaVisitor
Adjusts SharedResourcesManager to only call ProjectDeltaVisitor on
deltas for the shared reference points. This avoids a lot of overhead
otherwise caused by iterating and converting all non-shared changed
resources.

Adjusts ProjectDeltaVisitor to get passed a reference point object to
use. This skips having to guess which reference point the resources
belong to.

Removes ProjectDeltaVisitor.reset() as there now is one visitor per
reference point, meaning it does not have to get reset.

Adds the utility method createProjectReferencePointMapping which creates
a map of projects onto their contained reference points. This is useful
to quickly access all reference points for a specific project. This is
needed as we still filter by projects to check whether the resource
delta was created by the project being opened.