[INTERNAL][E] Update logic to set selections
Updates the logic in BaseResourceSelectionComposite for setting new
selections to work with the held list of checked resources.

Moves the old method to setSelectedResourcesInternal. Adds a new method
setSelectedResources that first reduces the given set of resources to
the contained base resources. This is done to ensure that our set of
held base resources does not contain any child resources, which would
result in Saros creating nested reference points.