[FIX][E] Use setSelectedResources(...) in ResourceSelectionComposite
uses setSelectedResources(...) in the logic creating new projects in
ResourceSelectionComposite. This ensures that the necessary listeners
are called and the internal state is updated correctly for the
additional selection.

Makes checkboxTreeViewer private in BaseResourceSelectionComposite as it
is now no longer accessed from outside the class. This ensures that the
model can't be updated incorrectly from the outside.