[INTERNAL][E] Adjust logic setting initial resource selection
Moves the addition of the selection listener before setting the
preselected resources to ensure that we are informed about the state
change.

Get list of displayed projects and use it directly to set the preset if
only one project is shown instead of using all shown resources.
Subsequently removes the method used to obtain all displayed resources.

Adjusts logic checking the initial input to also display the error
warning if no resources were selected by default.