[INTERNAL][E] Adjust logic setting initial resource selection
Moves the addition of the selection listener before setting the
preselected resources to ensure that we are informed about the state
change.

Removes the logic setting the project as selected if only one project
exists in the workspace and no preselection is present as it is no
longer useful with the new sharing model.

Adjusts logic checking the initial input to use the selection of the
resource page instead of the size of the preselection set as an
indicator of whether there is currently a selection. This is necessary
as the provided preselection might be completely illegal, meaning it
could still result in an empty selection.