[INTERNAL][E] Rework EnterProjectNamePage
Rewrites EnterProjectNamePage to better match the new resource sharing
model and the new ReferencePointOptionComposite design.

Updates the validation logic to match the new input fields. Trims the
input strings to avoid issues where values were seen as different due to
leading or trailing spaces that are dropped during the resource
conversion, leading to resource clashes.

Updates the shown error messages to make it clearer which reference
point composite they belong to.

Updates the error message logic to always display the error message of
the currently selected tab (if present). This makes it much easier for
the user to determine what the issue with the current input actually is.
Furthermore, if the error messages are updated and there is no error for
the current tab, the error message is only changed if the currently
displayed message is no longer valid. This was done to avoid potentially
confusing the user be unnecessarily switching the error message even
though both messages are not related to the current tab.

Updates the preselection logic. The new logic now always selects the
best guess, even if it is not a valid selection. This approach was
chosen as it forces the user to manually deal with potential conflicts
instead of just giving an easy alternative. This should hopefully give
the user a better understanding of what the outcome of the resource
negotiation will be.
Updates the preselection logic to also suggest using an existing project
if a shared reference point has the same name. This heuristic was chosen
as project names should be somewhat unique in a normal project.
Furthermore, sharing entire projects should be a common use case, so
making it as easy to do so should increase the usability of the
preselection logic.

The preselection logic will also be extended in a future PR by using the
workspace relative path on the sender side to provide the basis for a
good guess on how to map the reference point locally.

Updates the text displayed in the composite, the variable and method
names, and the javadoc to match the new resource sharing model.