[FIX][STF] Fix logic to use existing directory
Fixes the method to start a session on an existing directory by using
the correct label to look for.

Removes now unnecessary workaround deselecting the previously selected
radio button before selecting a new one. This logic was added to avoid a
bug which is only present in SWTBot versions <2.22.0.

Adds a sleep after switching the radio buttons. This sleep is necessary
to ensure that the internal listener reacting to the change event has
finished updating the UI. Otherwise, entering the text into the
directory path field fails as it is seen as read-only.

Adds constants for the labels to look for in the
AddReferencePointsToSessionWizard.