[FEATURE][CORE] Provide possible reference point representation options
Extends the AdditionalResourceData by adding possible resource
representations. These possible representation options are defined
through the name of a reference object (e.g. Eclipse project, IntelliJ
module, etc.) and the path of the reference point resource relative to
that object. This should hopefully allow the receiving side to make a
good suggestion on how to map the received reference point if its
representation already exists on the receiving side.

A list of such representations is offered as the representation does not
always have to be unique.

The class AbstractPossibleRepresentationProvider can be extended by the
specific IDE implementations which have to provide the logic to obtain
the possible representations for a given reference point.

The data can be read from the resource negotiation data using a utility
method of AbstractPossibleRepresentationProvider. This was added as the
data has to be flattened and escaped multiple times before adding it to
the additional resource data mapping, so the reversal of that process is
also done by the class.

Adds the utility class EscapeUtil that can be used to escape, unescape
and split strings for a given delimiter and escape character.