[INTERNAL][E] Optimize reference point lookup in SharedProjectDecorator
Optimizes the reference point lookup in SharedProjectDecorator by
holding a reference to the last used reference point. This is useful as
the decorations happens in in a depth-first search order along the tree
branches, meaning all resources of a shared reference point are
processed together.

This optimization should reduce the reference point lookup time to O(1)
instead of O(#ref-points) for most resources. This somewhat alleviates
the background activity at session start caused by Eclipse creating the
decorators for all shared resources, independently of whether they are
currently displayed or not.