[LOG][STF] Use log4j2 in STF
Added an additional logger to saros_log4j2.xml
that avoid swtbot trace messages. This config
is also used in Saros/E and Saros/I, but ignored,
because the package is not used. Making the whole
logic conditionally is much more complex, requires
more code and the benefit is very small.

Moved the log setup into the Activator to use
the bundle methods to determine the log directory.
Futhermore, the Activator is loaded before the
StartupSarosSTF class, thefore we make sure
that the logging is initialized before the first
logging call.